lib: DRY up U2F request authentication
This commit moves the common code between plain old U2F and DUO
implementations in to lib/u2f.go
