+++ lib/okta.go
@@ -364,56 +364,10 @@ func (o *OktaClient) u2fChallenge(oktaFactorID string) ([]byte, error) {
 		KeyHandle: o.UserAuth.Embedded.Factor.Profile.CredentialID,
 	}
 
-	allDevices := u2fhost.Devices()
-	// Filter only the devices that can be opened.
-	openDevices := []u2fhost.Device{}
-	for i, device := range allDevices {
-		err := device.Open()
-		if err == nil {
-			openDevices = append(openDevices, allDevices[i])
-			defer func(i int) {
-				allDevices[i].Close()
-			}(i)
-		}
+	response, err := authenticateU2FRequest(request, o.UserAuth.Embedded.Factor.Embedded.Challenge.TimeoutSeconds)
+	if err != nil {
+		return nil, err
 	}
-
-	authTimeout := o.UserAuth.Embedded.Factor.Embedded.Challenge.TimeoutSeconds
-	if authTimeout == 0 {
-		authTimeout = 20
-	}
-
-	var response *u2fhost.AuthenticateResponse
-	prompted := false
-	timeout := time.After(time.Second * time.Duration(authTimeout))
-	interval := time.NewTicker(time.Millisecond * 250)
-	defer interval.Stop()
-	for {
-		if response != nil {
-			break
-		}
-		select {
-		case <-timeout:
-			log.Infof("Failed to get authentication response after %d seconds", authTimeout)
-			break
-		case <-interval.C:
-			for _, device := range openDevices {
-				response, err = device.Authenticate(request)
-				if err == nil {
-					break
-				} else if _, ok := err.(*u2fhost.TestOfUserPresenceRequiredError); ok {
-					if !prompted {
-						fmt.Println("Touch the flashing U2F device to authenticate...")
-					}
-					prompted = true
-				} else {
-					return nil, err
-				}
-			}
-		}
-	}
-
-	log.Debugf("Client Data: %s", response.ClientData)
-	log.Debugf("Signature Data: %s", response.SignatureData)
 
 	payload, err := json.Marshal(OktaU2FResponse{
 		StateToken:    o.UserAuth.StateToken,
